requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
findiff==0.8.0
idna==2.9
imageio==2.8.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
mysql-connector-python==8.0.19
networkx==2.4
numpy==1.18.2
pandas==1.0.3
Pillow==7.0.0
pkg-resources==0.0.0
protobuf==3.6.1
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.16.2
scipy==1.4.1
selenium==3.141.0
six==1.14.0
soupsieve==2.0
TA-Lib==0.4.17
trendln==0.1.10
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
